crawler_rag/requirements.txt
# Navegació i scraping
playwright==1.43.0
requests

# Models de llenguatge i embeddings
openai==1.12.0
tiktoken==0.6.0
sentence-transformers==2.2.2
transformers==4.38.2

# Utilitats d'IA
google-generativeai==0.3.2  # Google Gemini
azure-core==1.30.1
google-adk

# Tractament de PDFs i HTML
pdfplumber
beautifulsoup4
httpx==0.27.2

# Connexió PostgreSQL
psycopg2-binary




# Utilitats addicionals
# Configuració i entorn
python-dotenv
textwrap3  # textwrap plus
python-dotenv

tabulate







pandas


